finished merge and refactoring
