+++ extension/chrome/settings/setup.ts
@@ -7,7 +7,7 @@ import { Url } from '../../js/common/core/common.js';
 import { ApiErr } from '../../js/common/api/shared/api-error.js';
 import { Assert } from '../../js/common/assert.js';
 import { Catch } from '../../js/common/platform/catch.js';
-import { Key, KeyInfoWithIdentity, KeyUtil } from '../../js/common/core/crypto/key.js';
+import { KeyInfoWithIdentity, KeyUtil } from '../../js/common/core/crypto/key.js';
 import { Gmail } from '../../js/common/api/email-provider/gmail/gmail.js';
 import { Google } from '../../js/common/api/email-provider/gmail/google.js';
 import { KeyImportUi } from '../../js/common/ui/key-import-ui.js';
@@ -31,8 +31,6 @@ import { KeyManager } from '../../js/common/api/key-server/key-manager.js';
 import { SetupWithEmailKeyManagerModule } from './setup/setup-key-manager-autogen.js';
 import { shouldPassPhraseBeHidden } from '../../js/common/ui/passphrase-ui.js';
 import Swal from 'sweetalert2';
-import { PassphraseStore } from '../../js/common/platform/store/passphrase-store.js';
-import { ContactStore } from '../../js/common/platform/store/contact-store.js';
 
 export interface SetupOptions {
   passphrase: string;
@@ -241,20 +239,6 @@ export class SetupView extends View {
 
   public finalizeSetup = async (): Promise<void> => {
     await AcctStore.set(this.acctEmail, { setup_date: Date.now(), setup_done: true, cryptup_enabled: true });
-  };
-
-  public saveKeysAndPassPhrase = async (prvs: Key[], options: SetupOptions) => {
-    for (const prv of prvs) {
-      await KeyStore.add(this.acctEmail, prv);
-      await PassphraseStore.set((options.passphrase_save && !this.clientConfiguration.forbidStoringPassPhrase()) ? 'local' : 'session',
-        this.acctEmail, { longid: KeyUtil.getPrimaryLongid(prv) }, options.passphrase);
-    }
-    const { sendAs } = await AcctStore.get(this.acctEmail, ['sendAs']);
-    const myOwnEmailsAddrs: string[] = [this.acctEmail].concat(Object.keys(sendAs!));
-    const { full_name: name } = await AcctStore.get(this.acctEmail, ['full_name']);
-    for (const email of myOwnEmailsAddrs) {
-      await ContactStore.update(undefined, email, { name, pubkey: KeyUtil.armor(await KeyUtil.asPublicKey(prvs[0])) });
-    }
   };
 
   public shouldSubmitPubkey = (checkboxSelector: string) => {
